% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{simple_theme}
\alias{simple_theme}
\title{Generates a simple theme that lies the focus on the data, not the graph}
\usage{
simple_theme()
}
\value{
a ggplot2 theme
}
\description{
Generates a simple theme that lies the focus on the data, not the graph
}
